agent/pyproject.toml
[project]
name = "roam-guide-agent"
version = "0.1.0"
description = "LiveKit voice agent for Roam tour guide powered by Gemini 2.5 Live"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[google,images]~=1.4",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
